requirements.txt
streamlit
fastapi
uvicorn
langchain
langchain-groq
langchain-community
chromadb
pandas
python-dotenv
uuid
streamlit
langchain
langchain-groq
langchain-community
chromadb
pandas
python-dotenv
fastapi
uvicorn
pydantic
